// bachelor-front-react-version-main/src/generated-sources/openapi/api.ts
export const OrderStatus = {
  AwaitingShipment: "AWAITING_SHIPMENT",
  ReadyForShipment: "READY_FOR_SHIPMENT",
  Shipped: "SHIPPED",
  AwaitingPickup: "AWAITING_PICKUP",
  Completed: "COMPLETED",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

/**
 *
 * @export
 * @interface OrderToFulfill
 */
export interface OrderToFulfill {
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  id?: string;
  /**
   *
   * @type {Array<ProductInOrder>}
   * @memberof OrderToFulfill
   */
  productList?: Array<ProductInOrder>;
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  fulfilledBy?: string;
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  orderedBy?: string;
  /**
   *
   * @type {ShippingDetails}
   * @memberof OrderToFulfill
   */
  shippingDetails?: ShippingDetails;
  /**
   *
   * @type {string}
   * @memberof OrderToFulfill
   */
  createdDate?: string;
  /**
   *
   * @type {OrderStatus}
   * @memberof OrderToFulfill
   */
  orderStatus?: OrderStatus;
}
// ...
export interface PatchOrderToFulfillProductStatusRequest {
  /**
   *
   * @type {string}
   * @memberof PatchOrderToFulfillProductStatusRequest
   */
  orderId: string;
  /**
   *
   * @type {string}
   * @memberof PatchOrderToFulfillProductStatusRequest
   */
  productId: string;
  /**
   *
   * @type {OrderStatus}
   * @memberof PatchOrderToFulfillProductStatusRequest
   */
  status: OrderStatus;
}
/**
 *
 * @export
 * @interface PatchOrderToFulfillStatusRequest
 */
export interface PatchOrderToFulfillStatusRequest {
  /**
   *
   * @type {string}
   * @memberof PatchOrderToFulfillStatusRequest
   */
  orderId: string;
  /**
   *
   * @type {OrderStatus}
   * @memberof PatchOrderToFulfillStatusRequest
   */
  status: OrderStatus;
}
// ...
export interface ProductInOrder {
  /**
   *
   * @type {string}
   * @memberof ProductInOrder
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof ProductInOrder
   */
  name?: string;
  /**
   *
   * @type {number}
   * @memberof ProductInOrder
   */
  price?: number;
  /**
   *
   * @type {string}
   * @memberof ProductInOrder
   */
  description?: string;
  /**
   *
   * @type {ProductInOrderMainImage}
   * @memberof ProductInOrder
   */
  mainImage?: ProductInOrderMainImage;
  /**
   *
   * @type {number}
   * @memberof ProductInOrder
   */
  amount?: number;
  /**
   *
   * @type {OrderStatus}
   * @memberof ProductInOrder
   */
  orderStatus?: OrderStatus;
}
/**
 *
 * @export
 * @interface ProductInOrderMainImage
 */
export interface ProductInOrderMainImage {
  /**
   *
   * @type {string}
   * @memberof ProductInOrderMainImage
   */
  imageKey?: string;
  /**
   *
   * @type {string}
   * @memberof ProductInOrderMainImage
   */
  imageUrl?: string;
}
// ...
export interface ShippingDetails {
  /**
   *
   * @type {string}
   * @memberof ShippingDetails
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof ShippingDetails
   */
  city?: string;
  /**
   *
   * @type {string}
   * @memberof ShippingDetails
   */
  country?: string;
  /**
   *
   * @type {string}
   * @memberof ShippingDetails
   */
  address?: string;
  /**
   *
   * @type {string}
   * @memberof ShippingDetails
   */
  postalCode?: string;
}